requirements.txt
beautifulsoup4==4.9.3
hupper==1.10.2
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
PasteDeploy==2.1.1
plaster==1.0
plaster-pastedeploy==0.7
psycopg2==2.8.6
Pygments==2.8.0
pyramid==1.10.7
pyramid-debugtoolbar==4.9
pyramid-jinja2==2.8
pyramid-mako==1.1.0
-e git+https://github.com/EdwyReed/Pyramid-ToDo.git@7f5cfa482d00516aa3bc6f24c94eea732507e89d#egg=PyramidToDo
python-dotenv==0.15.0
repoze.lru==0.7
six==1.15.0
soupsieve==2.2
translationstring==1.4
venusian==3.0.0
waitress==1.4.4
WebOb==1.8.7
WebTest==2.0.35
zope.deprecation==4.4.0
zope.interface==5.2.0
